% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_moodle_web_questions.R
\name{add_moodle_quiz_questions_2}
\alias{add_moodle_quiz_questions_2}
\title{experimental}
\usage{
add_moodle_quiz_questions_2(
  question_source_files,
  tmp_dir = here::here(),
  tmp_prefix = NULL,
  include_solution = FALSE,
  include_metadata = FALSE,
  question_number_fmt = "## Question \%1$0.2d: \%2$s",
  solution_section_fmt = "### Solution"
)
}
\description{
experimental
}
