% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_assignment_doc.R
\name{build_answer_key}
\alias{build_answer_key}
\title{Create an answer key for a moodle quiz-question based assignment.}
\usage{
build_answer_key(
  doc_source_file,
  question_source_files,
  key_target_file = NULL,
  include_solution = FALSE,
  include_metadata_key = FALSE,
  question_number_fmt = "## Question \%1$0.2d: \%2$s",
  solution_section_fmt = "### Solution"
)
}
\description{
Create an answer key for a moodle quiz-question based assignment.
}
