% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_assignment_doc.R
\name{build_assignment_doc}
\alias{build_assignment_doc}
\title{Create an answer key for a moodle quiz-question based assignment.}
\usage{
build_assignment_doc(
  doc_source_file,
  question_source_files,
  doc_target_file = NULL,
  key_target_file = NULL,
  include_solution = FALSE,
  include_metadata = FALSE,
  include_metadata_key = FALSE,
  question_number_fmt = "## Question \%1$0.2d: \%2$s",
  solution_section_fmt = "### Solution"
)
}
\arguments{
\item{doc_source_file}{a}

\item{question_source_files}{aaa}

\item{doc_target_file}{blah}

\item{key_target_file}{blah}

\item{include_solution}{blah}

\item{include_metadata}{blah}

\item{include_metadata_key}{blah}

\item{question_number_fmt}{blah}

\item{solution_section_fmt}{blah}
}
\description{
Create an answer key for a moodle quiz-question based assignment.
}
