% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_moodle_web_questions.R
\name{combine_moodle_quiz_question_source}
\alias{combine_moodle_quiz_question_source}
\title{Combine questions from mutliple moodle quiz question source files}
\usage{
combine_moodle_quiz_question_source(
  question_source_files,
  include_solution = FALSE,
  include_metadata = FALSE,
  question_number_fmt = "## Question \%1$0.2d: \%2$s",
  solution_section_fmt = "### Solution"
)
}
\description{
Combine questions from mutliple moodle quiz question source files
}
