% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_file.R
\name{find_file}
\alias{find_file}
\title{Find a file}
\usage{
find_file(
  filename,
  search_path = NULL,
  return_all = FALSE,
  duplicated_files_error = FALSE,
  verbose = FALSE,
  directory = FALSE,
  extension = NULL,
  error_if_none = TRUE
)
}
\arguments{
\item{filename}{Name of the file to search for}

\item{search_path}{Optional path within which to search}

\item{return_all}{Return all matches?}

\item{duplicated_files_error}{Throw an error if more than one matching file found?}

\item{verbose}{Print messages?}

\item{directory}{Search for a directory instead of a file?}

\item{extension}{Search for files with this extension only?}

\item{error_if_none}{Throw an error if no matches are found?  If FALSE, the function returns a NULL value.}
}
\value{
The absolute path to the file, if it was found.
}
\description{
Locate a file somewhere within the directory structure of an RProject.
}
