% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_moodle_web_questions.R
\name{format_moodle_question_source}
\alias{format_moodle_question_source}
\title{Format a moodle question source file}
\usage{
format_moodle_question_source(
  lines_i,
  include_question = TRUE,
  include_solution = FALSE,
  include_metadata = FALSE
)
}
\arguments{
\item{lines_i}{a character vector containing the lines of a Rmd source file.}

\item{include_question}{boolean: whether nor not to include the 'Solution' section of the source files}

\item{include_solution}{boolean: whether nor not to include the 'Solution' section of the source files}

\item{include_metadata}{boolean: whether or not to include the 'Meta-information' section of the source files.}

\item{insert_line_breaks_in_metadata}{}
}
\description{
Prepare the text of a moodle question source file by including or excluding
the "Solution" and "Meta-information" sections.
}
