% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_moodle_web_questions.R
\name{format_moodle_web_questions}
\alias{format_moodle_web_questions}
\title{Format moodle question source file}
\usage{
format_moodle_web_questions(
  question_source_files,
  include_solution = FALSE,
  include_metadata = FALSE
)
}
\arguments{
\item{question_source_files}{a character vector of filenames of source files to be parsed and formatted}

\item{include_solution}{boolean: whether nor not to include the 'Solution' section of the source files}

\item{include_metadata}{boolean: whether or not to include the 'Meta-information' section of the source files.}
}
\value{
source code lines
}
\description{
Format moodle question source file for insertion in a html doc such as
an assignment file, answer key, etc.
}
