% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_question_section.R
\name{get_f_lines}
\alias{get_f_lines}
\title{Retrieve the contents of a source .Rmd file}
\usage{
get_f_lines(f_name = NULL, f_lines = NULL)
}
\arguments{
\item{f_name}{A character representing the path to a .Rmd source file.}

\item{f_lines}{A character vector with the lines of the .Rmd source file.}
}
\description{
Checks to see if the file has already been read.
If not, returns a character vector of the source file lines.
If so, simply returns the already read source lines.
}
