% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_moodle_web_questions.R
\name{get_moodle_question_section_line_indices}
\alias{get_moodle_question_section_line_indices}
\title{Get the line indices for a section}
\usage{
get_moodle_question_section_line_indices(
  rmd_source_lines,
  section_name = "Meta-information",
  section_delimiter = "====="
)
}
\arguments{
\item{rmd_source_lines}{A character vector with the lines of the .Rmd source file.}

\item{section_name}{Name of section, e.g. "Question", "Solution", "Meta-information"}

\item{section_delimiter}{The section break delimiter to search for}
}
\description{
Finds the line indices of a section in an .Rmd file formatted
for use with the 'exams' package.
}
