% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_question_section.R
\name{get_next_index}
\alias{get_next_index}
\title{naaaaa na na na na na na "utility function!"}
\usage{
get_next_index(start_indices, delim_indices, f_lines)
}
\arguments{
\item{start_indices}{blah}

\item{delim_indices}{blah}

\item{f_lines}{A character vector with the lines of the .Rmd source file.}
}
\description{
naaaaa na na na na na na "utility function!"
}
