% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_question_section.R
\name{get_post_header}
\alias{get_post_header}
\title{Get the line indices for a section}
\usage{
get_post_header(
  f_name = NULL,
  f_lines = NULL,
  header = "Question\\\\s{0,}$",
  yaml_header_delim = "---"
)
}
\arguments{
\item{f_name}{A character representing the path to a .Rmd source file.}

\item{f_lines}{A character vector with the lines of the .Rmd source file.}

\item{header}{Name of section, e.g. "Question", "Solution", "Meta-information"}

\item{yaml_header_delim}{blah}
}
\description{
Finds the line indices of a section in an .Rmd file formatted
for use with the 'exams' package.
}
