% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_question_markup.R
\name{get_question_markup}
\alias{get_question_markup}
\title{Get markup for a question?}
\usage{
get_question_markup(
  q_files,
  q_fmt = "## Question \%s",
  q_header = "Question\\\\s{0,}$",
  delim = "^={3,}\\\\s{0,}$",
  a_list_fmt = "answerlist"
)
}
\arguments{
\item{q_files}{blah}

\item{q_fmt}{blah}

\item{q_header}{blah}

\item{delim}{blah}

\item{a_list_fmt}{blah}
}
\description{
Get markup for a question?
}
