% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_question_section.R
\name{get_question_section_indices}
\alias{get_question_section_indices}
\title{Get the line indices for a section}
\usage{
get_question_section_indices(
  f_name = NULL,
  f_lines = NULL,
  header = "Question\\\\s{0,}$",
  delim = "^={3,}\\\\s{0,}$",
  end_header = NULL
)
}
\arguments{
\item{f_name}{A character representing the path to a .Rmd source file.}

\item{f_lines}{A character vector with the lines of the .Rmd source file.}

\item{header}{Name of section, e.g. "Question", "Solution", "Meta-information"}

\item{delim}{The section break delimiter to search for.  By default
searches for three or more consecutive equals symbols.}

\item{end_header}{Name of section, e.g. "Question", "Solution", "Meta-information".  If Null, the default, the function looks for the ending delimiter for the section specified by 'header'.}
}
\description{
Finds the line indices of a section in an .Rmd file formatted
for use with the 'exams' package.
}
