% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rmd_chunks.R
\name{get_rmd_chunk_indices}
\alias{get_rmd_chunk_indices}
\title{Get the yaml header from an Rmd file}
\usage{
get_rmd_chunk_indices(filename, file_lines = NULL)
}
\arguments{
\item{filename}{the name of the Rmd file to read}

\item{file_lines}{a character vector containing the lines of a source Rmd file.}
}
\description{
Get the yaml header from an Rmd file
}
