% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rmd_header.R
\name{get_rmd_header_attr}
\alias{get_rmd_header_attr}
\title{desxription}
\usage{
get_rmd_header_attr(filename, file_lines = NULL, header_prefix = "title:")
}
\arguments{
\item{filename}{the name of the Rmd file to read}

\item{file_lines}{a character vector containing the lines of a source Rmd file.}

\item{header_prefix}{blah}
}
\description{
desxription
}
