% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_moodle_web_questions.R
\name{get_sec}
\alias{get_sec}
\title{get a section from a rmd file}
\usage{
get_sec(headr, lines)
}
\arguments{
\item{headr}{name of the header}

\item{lines}{how many lines to get}
}
\description{
get a section from a rmd file
}
