% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_line_breaks.R
\name{insert_line_breaks}
\alias{insert_line_breaks}
\title{Insert newline characters between lines}
\usage{
insert_line_breaks(input_lines, lb = "\\n")
}
\arguments{
\item{input_lines}{lines in-between which to insert newline characters}

\item{lb}{new line character to insert}
}
\description{
Adds a new line character in between lines contained in a character vector
}
