% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_html_dropdown.R
\name{make_html_dropdown}
\alias{make_html_dropdown}
\title{Create html code for a dropdown menu.}
\usage{
make_html_dropdown(
  options,
  fmt_select = "<select>\\n\%s\\n</select>",
  fmt_option = "<option>\%s</option>",
  schoice_q = "select the best answer",
  mchoice_q = "select the best answer or answers",
  mchoice = FALSE,
  cat = TRUE
)
}
\arguments{
\item{options}{a vector containing the elements to include int he dropdown}

\item{fmt_select}{the format for the html select element}

\item{fmt_option}{the format for the html option elemment}

\item{schoice_q}{the prompt text to display for multiple-choice}

\item{mchoice}{whether or not to display the multiple-choice question description}

\item{cat}{whether to cat() the results}
}
\description{
Create html code for a dropdown menu.
}
