% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qn.R
\name{qn}
\alias{qn}
\title{Format a questionnumber and point count for RMarkdown}
\usage{
qn(
  n,
  pts = 1,
  q_string = "Q",
  pt_string = " pt.",
  pts_string = " pts.",
  pt_brackets = c("(", ")"),
  end_string = ":"
)
}
\arguments{
\item{n}{Question number}

\item{pts}{Number of points}

\item{q_string}{Question abbreviation to show}

\item{pt_string}{Points text for a 1-point question}

\item{pts_string}{Points text for a question worth more than 2 pts}

\item{pt_brackets}{Brackets to surround the number of poitns}

\item{end_string}{Ending string}
}
\value{
Markdown formatted text for the question number and point count.
}
\description{
Format a questionnumber and point count for RMarkdown
}
