% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_page.R
\name{render_page}
\alias{render_page}
\title{Wrapper for render
Includes argument to check for specific version of pandoc or higher}
\usage{
render_page(filename_in, filename_out, dir_out, pandoc_version = "2.11")
}
\arguments{
\item{filename_in}{Name of input file}

\item{filename_out}{Name of output file.  Given without extension: the extension will be added according to the output format specified in the header.}

\item{dir_out}{output directory}

\item{pandoc_version}{Required version of pandoc.  If this version (or higher) is not found, the page will not render.}
}
\description{
Wrapper for render
Includes argument to check for specific version of pandoc or higher
}
