% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_lines.R
\name{source_lines}
\alias{source_lines}
\title{Source specific lines in an R file}
\usage{
source_lines(file_lines, lines)
}
\arguments{
\item{file_lines}{character vector containing lines read from a source file}

\item{lines}{numeric vector of lines to source in \code{file}.}
}
\description{
Source specific lines in an R file
}
