% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rmd_chunks.R
\name{source_rmd_chunks}
\alias{source_rmd_chunks}
\title{description}
\usage{
source_rmd_chunks(filename, file_lines = NULL, chunk_names = NULL)
}
\arguments{
\item{filename}{the name of the Rmd file to read}

\item{file_lines}{a character vector containing the lines of a source Rmd file.}

\item{chunk_names}{blah}
}
\description{
description
}
