% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_html_dropdown.R
\name{sub_cloze_answer}
\alias{sub_cloze_answer}
\title{Create substitute text for a cloze answer}
\usage{
sub_cloze_answer(
  ans_num,
  check_var,
  sub_text = "<span>______</span>",
  cloze_fmt = "\\\\#\\\\#ANSWER\%s\\\\#\\\\#",
  cat = FALSE
)
}
\arguments{
\item{ans_num}{cloze answer number}

\item{check_var}{if an object with this name exists, the substitute text will be returned, otherwise the cloze answer is returned}

\item{sub_text}{text to substitute for the cloze code}

\item{cloze_fmt}{format for the cloze code}

\item{cat}{whether or not to cat() the output}
}
\description{
Create substitute text for a cloze answer
}
