% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rmd_header.R
\name{substitute_rmd_header_attr}
\alias{substitute_rmd_header_attr}
\title{Swap the value of an Rmd header attribute}
\usage{
substitute_rmd_header_attr(header_lines, attr_prefix = "title:", new_attr_val)
}
\arguments{
\item{header_lines}{A character vector with the lines of the header}

\item{attr_prefix}{The name of the attribute (with ending colon)}

\item{new_attr_val}{The value to substitute}
}
\description{
Swap the value of an Rmd header attribute
}
