% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_moodle_web_questions.R
\name{build_assignment_doc}
\alias{build_assignment_doc}
\title{Create an answer key for a moodle quiz-question based assignment.}
\usage{
build_assignment_doc(
  assignment_source_file,
  question_source_files,
  assignment_render_file = NULL,
  key_render_file = NULL,
  include_solution = FALSE,
  include_metadata = FALSE,
  question_number_fmt = "## Question \%1$0.2d: \%2$s",
  solution_section_fmt = "### Solution"
)
}
\description{
Create an answer key for a moodle quiz-question based assignment.
}
