% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_moodle_xml_questions.R
\name{build_moodle_xml_questions}
\alias{build_moodle_xml_questions}
\title{Wrapper for exams2moodle}
\usage{
build_moodle_xml_questions(
  question_source_files,
  assignment_name,
  xml_output_path,
  exam_mchoice = list(shuffle = TRUE),
  exam_schoice = list(shuffle = TRUE),
  xml_output_filename = NULL
)
}
\arguments{
\item{question_source_files}{blah}

\item{assignment_name}{blah}

\item{xml_output_path}{blah}

\item{exam_mchoice}{list(shuffle = TRUE)}

\item{exam_schoice}{list(shuffle = TRUE)}

\item{xml_output_filename}{= NULL}
}
\description{
Wrapper for exams2moodle
}
