% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_moodle_web_questions.R
\name{format_moodle_web_questions}
\alias{format_moodle_web_questions}
\title{Format moodle question source file}
\usage{
format_moodle_web_questions(
  f_names,
  q_header = "Question\\\\s{0,}$",
  delim = "^={3,}\\\\s{0,}$",
  cloze_regex = "\\\\\\\\#\\\\\\\\#ANSWER[0-9]*\\\\\\\\#\\\\\\\\#",
  cloze_replacement = "________",
  rm_css_chunk_name = TRUE,
  include_question = TRUE,
  include_solution = FALSE,
  include_metadata = FALSE,
  soln_header = "Solution\\\\s{0,}$",
  meta_header = "Meta-information\\\\s{0,}$",
  q_ext = ".Rmd"
)
}
\arguments{
\item{f_names}{a}

\item{q_header}{a}

\item{delim}{a}

\item{cloze_regex}{blah}

\item{cloze_replacement}{= "________",}

\item{rm_css_chunk_name}{= TRUE,}

\item{include_question}{= TRUE,}

\item{include_solution}{= FALSE,}

\item{include_metadata}{= FALSE,}

\item{soln_header}{blah}

\item{meta_header}{blah}

\item{q_ext}{blah}
}
\value{
source code lines
}
\description{
Format moodle question source file for insertion in a html doc such as
an assignment file, answer key, etc.
}
