% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_assignment_paths.R
\name{gather_assignment_paths}
\alias{gather_assignment_paths}
\title{Create an answer key for a moodle quiz-question based assignment.}
\usage{
gather_assignment_paths(
  week_n,
  assignment_prefix,
  out_dir,
  moodle_xml_out_dir = here("docs", "moodle_quiz_questions"),
  out_fmt = "\%s_\%0.2d",
  key_out_dir = here("docs", "answer_keys"),
  key_suffix = "answer_key",
  source_dir_fmt = "\%s_\%0.2d_",
  source_file_fmt = "\%s_\%0.2d",
  source_search_path = NULL,
  moodle_question_source_subdir = "moodle",
  moodle_question_prefix = "Q_",
  key_file_fmt = "\%s_\%0.2d_answer_key",
  moodle_xml_out_fmt = "\%s_\%0.2d_moodle_questions",
  doc_source_ext = ".Rmd",
  question_source_ext = ".Rmd"
)
}
\arguments{
\item{week_n}{blah}

\item{assignment_prefix}{blah}

\item{out_dir}{blah}

\item{moodle_xml_out_dir}{blah}

\item{out_fmt}{blah}

\item{key_out_dir}{blah}

\item{key_suffix}{blah}

\item{source_dir_fmt}{blah}

\item{source_file_fmt}{blah}

\item{source_search_path}{bl;ah}

\item{moodle_question_source_subdir}{blah}

\item{moodle_question_prefix}{blah}

\item{key_file_fmt}{blah}

\item{moodle_xml_out_fmt}{blah}

\item{doc_source_ext}{bklah}

\item{question_source_ext}{blah}

\item{answer_key}{blah}
}
\description{
Create an answer key for a moodle quiz-question based assignment.
}
