% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_question_section.R
\name{get_question_section}
\alias{get_question_section}
\title{Retrieve lines from a source .Rmd file, formatted as a Moodle quiz question from package 'exams'}
\usage{
get_question_section(
  f_name = NULL,
  f_lines = NULL,
  header = "Question\\\\s{0,}$",
  delim = "^={3,}\\\\s{0,}$",
  cloze_regex = "\\\\\\\\#\\\\\\\\#ANSWER[0-9]*\\\\\\\\#\\\\\\\\#",
  cloze_replacement = "________",
  rm_css_chunk_name = TRUE,
  end_header = NULL
)
}
\arguments{
\item{f_name}{A character representing the path to a .Rmd source file.}

\item{f_lines}{A character vector with the lines of the .Rmd source file.}

\item{header}{Name of section, e.g. "Question", "Solution", "Meta-information"}

\item{delim}{The section break delimiter to search for.  By default
searches for three or more consecutive equals symbols.}

\item{cloze_regex}{blah}

\item{cloze_replacement}{blah}

\item{rm_css_chunk_name}{blah TRUE}

\item{end_header}{blah NULL}
}
\description{
f_name = NULL, f_lines = NULL,
}
