% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rmd_chunks.R
\name{get_rmd_chunks}
\alias{get_rmd_chunks}
\title{Get all the chunk names and the start/end line numbers for the corresponding source code.}
\usage{
get_rmd_chunks(filename = NULL, file_lines = NULL)
}
\arguments{
\item{filename}{the name of the Rmd file to read}

\item{file_lines}{a character vector containing the lines of a source Rmd file.}
}
\description{
Get all the chunk names and the start/end line numbers for the corresponding source code.
}
